requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
arrow==1.3.0
atomicwrites==1.4.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cssselect==1.3.0
cycler==0.12.1
docstring-to-markdown==0.16
et_xmlfile==2.0.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
jsonpointer==2.1
kiwisolver==1.4.7
lxml==5.3.2
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
matplotlib==3.9.2
multitasking==0.0.11
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.0.2
openpyxl==3.2.0b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
proto-plus==1.26.1
protobuf>=4.26.1,<6.0dev
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.11.1
pydantic_core==2.33.0
pyls-spyder==0.4.0
pyparsing==3.2.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.16
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rsa==4.9
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
webcolors==24.11.1
Werkzeug==3.1.3
wordcloud==1.9.4
yfinance==0.2.55
zipp==3.21.0
flask-dance[sqla]>=7.0.0
